requirements.txt
Django==1.7
MySQL-python==1.2.5
beautifulsoup4==4.3.2
django-tastypie==0.12.0
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.4.3
six==1.8.0
wsgiref==0.1.2
